support varchar -> scala enum, and pg enum -> scala enum
