+++ magnum-pg/src/test/scala/Color.scala
@@ -1,0 +1,4 @@
+import com.augustnagro.magnum.DbCodec
+
+enum Color derives DbCodec:
+  case Red, Green, Blue
